Tickets and analytics

+++ FrontEnd/src/pages/admin/AdminAssessments/AdminAssessments.jsx
@@ -944,6 +944,16 @@ const AdminAssessments = () => {
   };
 
   const handleDeleteAssessment = async (id) => {
+    const confirmed = await confirm({
+      title: `Are you sure you want to delete this Assessment?`,
+      confirmText: 'Delete',
+      cancelText: 'Cancel',
+      type: 'danger', // or 'warning', 'info'
+      showCheckbox: true,
+      checkboxLabel: 'I understand that the data cannot be retrieved after deleting.',
+      note: 'Associated items will be removed.',
+    });
+    if (!confirmed) return;
     try {
       const res = await dispatch(deleteAssessment(id));
       if (deleteAssessment.fulfilled.match(res)) {
